Updated Assertion Error in _runAssertionOnResults =======================================================
There's a check which confirms that the length of the tolerance
specified values is the same as computed vals (which was
asserted to be the same as the reference vals). This check
must be true, but in the case where a tolerance is specifed
as a single tuple a list is generated from that tuple equiv
in length to the reference vals. Therefore the user would
have specified a single tuple (which isn't the same length
as the reference values). Therefore the error message was
potentially confusing and was updated