statement
The eval statement allows a user to specify some combination
of ABAQUS historical data. The user specifies, in a normal fashion,
dictionaries definining how to interrogate the abaqus data with
a dictionary in the identifier list. The difference being that
an extra key value pair is provided *av_id* . This is a unique
identifier of the identifying object so that it may be referenced in
either the x_eval_statement or y_eval_statement (additional key, vals
in an element of results list).

x and y eval statements are strings which define a valid
python syntax equation for somehow combining the xy hist
data returned by the identifier. Internally a dictionary
d is defined mapping labels -> xydata.

Therefore an eval statement would be written as follows:

d['label1'] + d['label2']