Added ability to specifying an eval statement for tabular data input. Useful for use cases such as combining multiple RFs to create a stress for comparison (#2)
The eval statement allows a user to specify some combination
of ABAQUS historical data. The user specifies, in a normal fashion,
dictionaries definining how to interrogate the abaqus data with
a dictionary in the identifier list. The difference being that
an extra key value pair is provided *av_id* . This is a unique
identifier of the identifying object so that it may be referenced in
either the x_eval_statement or y_eval_statement (additional key, vals
in an element of results list).

x and y eval statements are strings which define a valid
python syntax equation for somehow combining the xy hist
data returned by the identifier. Internally a dictionary
d is defined mapping labels -> xydata.

Therefore an eval statement would be written as follows:

d['label1'] + d['label2']

* 1. Updated test case to demonstrate arithmetic summation of results lists within tabular
2. Updated readme to have example
3. Renamed av_id to label
4. Changed underscore notation of new variables to be camelcase
   (more alignment with rest of project)

* # Small corrections to readme

* # Added ability to use eval statements in all result types

* removed debug statement