New test type: xy_infl_pt_bilinear
A variation of test type xy_infl_pt in which an x-y data set is split 
into two consecutive sets and linear curve fits are performed on each. 
The sum of the residuals of the two fittings are summed, and the 
intersection of the best fitting pair is defined as the inflection 
point. This test type performs well for piecewise linear histories which 
may include some higher frequency vibrations. The xy history is 
resampled to prior to splitting so as to equally weight the summed 
errors.